fix: refactor move ItemGSTDetails to taxes controller
